DAV2 - 0013
Added several statistics methods and changed the activation email domain to truesight.site

There are two pools of records used to get the averages.
To get the averages from all the records that have been entered in the application, use the endpoint:

getaverage/
Method: GET

If you only want to get the averages from the model training set, use this endpoint:
getaveragebase/
Method: GET

There are also the GET methods used to get new averages. These are stored in a table and the last record is called with the use of the methods above.
calculateaverage/
calculateaveragebase/

Didn't do the grouping of predictions by massivePredictionId because every massive prediction is composed of singular smaller predictions. The best prediction isn't the whole blob of predictions, buy a singular one inside of the set.